change to usize

+++ src/amm/uniswap_v2/factory/pair_addresses_batch_request.rs
@@ -14,8 +14,8 @@ use super::{contracts::IGetUniswapV2PairsBatchRequest, UniswapV2Factory};
 impl UniswapV2Factory {
     async fn get_pair_addresses_from_factory_batch<M: Middleware>(
         &self,
-        start: u64,
-        end: u64,
+        start: usize,
+        end: usize,
         middleware: Arc<M>,
         progress_bar: Option<Arc<Mutex<ProgressBar>>>,
     ) -> Vec<H160> {
@@ -58,13 +58,13 @@ impl UniswapV2Factory {
 
     pub async fn get_pair_addresses_from_factory<'a, M: Middleware + 'a>(
         &self,
-        start: u64,
-        end: u64,
+        start: usize,
+        end: usize,
         step: usize,
         middleware: Arc<M>,
     ) -> Vec<H160> {
         let batch_func =
-            |start: u64, end: u64, middleware: Arc<M>, pb: Option<Arc<Mutex<ProgressBar>>>| {
+            |start: usize, end: usize, middleware: Arc<M>, pb: Option<Arc<Mutex<ProgressBar>>>| {
                 self.get_pair_addresses_from_factory_batch(start, end, middleware.clone(), pb)
             };
         println!(
